requirements.txt
Django
djangorestframework
psycopg2-binary
django-modeltranslation
django-filter
unidecode